Conflicts:
	Octokit.Reactive/Octokit.Reactive.csproj
	Octokit.Reactive/packages.config
	Octokit/Models/Response/PullRequest.cs
	Octokit/Models/Response/Repository.cs
	Octokit/Octokit.csproj
	Octokit/Properties/AssemblyInfo.cs

+++ Octokit.Reactive/Clients/IObservableDeploymentsClient.cs
@@ -21,7 +21,7 @@ namespace Octokit.Reactive
         /// <param name="owner">The owner of the repository</param>
         /// <param name="name">The name of the repository</param>
         IObservable<Deployment> GetAll(string owner, string name);
-        
+
         /// <summary>
         /// Gets all the deployments for the specified repository. Any user with pull access
         /// to a repository can view deployments.
@@ -30,7 +30,7 @@ namespace Octokit.Reactive
         /// http://developer.github.com/v3/repos/deployments/#list-deployments
         /// </remarks>
         /// <param name="repositoryId">The Id of the repository</param>
-        IObservable<Deployment> GetAll(int repositoryId);
+        IObservable<Deployment> GetAll(long repositoryId);
 
         /// <summary>
         /// Gets all the deployments for the specified repository. Any user with pull access
@@ -43,7 +43,7 @@ namespace Octokit.Reactive
         /// <param name="name">The name of the repository</param>
         /// <param name="options">Options for changing the API response</param>
         IObservable<Deployment> GetAll(string owner, string name, ApiOptions options);
-        
+
         /// <summary>
         /// Gets all the deployments for the specified repository. Any user with pull access
         /// to a repository can view deployments.
@@ -53,7 +53,7 @@ namespace Octokit.Reactive
         /// </remarks>
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="options">Options for changing the API response</param>
-        IObservable<Deployment> GetAll(int repositoryId, ApiOptions options);
+        IObservable<Deployment> GetAll(long repositoryId, ApiOptions options);
 
         /// <summary>
         /// Creates a new deployment for the specified repository.
@@ -76,7 +76,7 @@ namespace Octokit.Reactive
         /// </remarks>
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="newDeployment">A <see cref="NewDeployment"/> instance describing the new deployment to create</param>
-        IObservable<Deployment> Create(int repositoryId, NewDeployment newDeployment);
+        IObservable<Deployment> Create(long repositoryId, NewDeployment newDeployment);
 
         /// <summary>
         /// Client for managing deployment status.
