+++ Octokit.Tests.Integration/Clients/RepositoryPagesClientTests.cs
@@ -10,7 +10,7 @@ public class RepositoryPagesClientTests
         readonly IRepositoryPagesClient _repositoryPagesClient;
         const string owner = "octokit";
         const string name = "octokit.net";
-        const int repositoryId = 7528679;
+        const long repositoryId = 7528679;
 
         public TheGetMethod()
         {
@@ -18,14 +18,14 @@ public class RepositoryPagesClientTests
             _repositoryPagesClient = github.Repository.Page;
         }
 
-        [IntegrationTest(Skip= "These tests require repository admin rights - see https://github.com/octokit/octokit.net/issues/1263 for discussion")]
+        [IntegrationTest(Skip = "These tests require repository admin rights - see https://github.com/octokit/octokit.net/issues/1263 for discussion")]
         public async Task ReturnsMetadata()
         {
             var data = await _repositoryPagesClient.Get(owner, name);
             Assert.Equal("https://api.github.com/repos/octokit/octokit.net/pages", data.Url);
         }
 
-        [IntegrationTest(Skip= "These tests require repository admin rights - see https://github.com/octokit/octokit.net/issues/1263 for discussion")]
+        [IntegrationTest(Skip = "These tests require repository admin rights - see https://github.com/octokit/octokit.net/issues/1263 for discussion")]
         public async Task ReturnsMetadataWithRepositoryId()
         {
             var data = await _repositoryPagesClient.Get(repositoryId);
